L3 layout: TransistorList int* → ushort* (697KB → 350KB; +1.9% avg)
The biggest hot array in L3, by far. Stores flattened node IDs (gate /
c1 / c2 / other) with 0-sentinel terminators between sub-lists. Node
IDs are <65K (full_palette: 14,723 nodes) so ushort fits. Halves the
working set — particularly L2 footprint, which is where BFS inner
loops fault when walking large group chains.

Bit-identical (checksum 0x933ABE7915AC18BE on full_palette.nes 50K hc).

A/B bench (on top of byte _inGroup + ushort _groupBuf):
  prior state:           best 38,591  avg 37,874  hc/s
  TransistorList ushort: best 39,056  avg 38,587  hc/s
                         → +1.2% best, +1.9% avg

Cumulative L3-layout series vs original aot-codegen baseline:
                        best       avg
  original (all int*):  37,928    36,923
  byte _inGroup:        38,534    38,018  (+3.0% avg)
  + ushort _groupBuf:   38,591    37,874  (+2.6% avg)
  + ushort TList:       39,056    38,587  (+4.5% avg)  ← here

Change pattern: declaration + alloc + 11 reader sites (`int* p =
TransistorList + ...` → `ushort* p = ...`). All reads use `*p++`
which auto-widens ushort → int. Writes use (ushort) cast.

SnapshotExporter writes (int)TransistorList[i] to preserve the v4
wire format — Rust port reads int-per-entry, unchanged.

Files: WireCore.cs (decl + alloc), Group/GroupSimd/Recalc/FastPath/
Ir/Levelize.cs (reader walks), SnapshotExporter.cs (widen on export).

+++ src/AprVisual/Sim/WireCore.FastPath.cs
@@ -80,12 +80,12 @@ namespace AprVisual.Sim
 
             if (ns.TlistC1gnd != 0)
             {
-                int* p = TransistorList + ns.TlistC1gnd;
+                ushort* p = TransistorList + ns.TlistC1gnd;
                 while (*p != 0) { if (NodeStates[*p++] != 0) { f |= NodeFlags.Gnd; break; } }   // any ON path to GND ⇒ pulled low
             }
             if (ns.TlistC1pwr != 0)
             {
-                int* p = TransistorList + ns.TlistC1pwr;
+                ushort* p = TransistorList + ns.TlistC1pwr;
                 while (*p != 0) { if (NodeStates[*p++] != 0) { f |= NodeFlags.Pwr; break; } }   // any ON path to VCC ⇒ pulled high
             }
